v8_hoshilm_submission/hoshilm_kr/requirements.txt
torch>=2.1
pyyaml>=6.0
sentencepiece>=0.1.99
matplotlib>=3.7
